CAP-35: Minimal changes required to build ClawbackOp (#1)
### What
Add the minimal changes required to build a Clawback operation for including in a transaction.

### Why
To experiment with Clawback in .NET applications.

### Notes
The diff is larger than it needs to be because the xdrgen tool is creating immaterial changes to the generated files. I've done all I can for the moment to address those and there are upstream issues discussing it (elucidsoft/dotnet-stellar-sdk#302 elucidsoft/dotnet-stellar-sdk#303). This issue doesn't impact our ability to experiment with this change.